Restruture project folders
