+++ app/_layout.tsx
@@ -1,7 +1,7 @@
 // app/_layout.tsx
 import { Slot } from 'expo-router';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
-import { ThemeProvider } from '../context/ThemeContext'; // ✅ Make sure path is correct
+import { ThemeProvider } from '../src/context/ThemeContext';
 
 export default function RootLayout() {
   return (
