+++ package.json
@@ -17,29 +17,30 @@
     "@react-navigation/elements": "^2.3.8",
     "@react-navigation/native": "^7.1.9",
     "@react-navigation/stack": "^7.3.2",
+    "axios": "^1.10.0",
     "date-fns": "^4.1.0",
-    "expo": "~53.0.9",
+    "expo": "53.0.12",
     "expo-apple-authentication": "~7.2.4",
     "expo-application": "^6.1.4",
     "expo-auth-session": "~6.2.0",
-    "expo-blur": "~14.1.4",
+    "expo-blur": "~14.1.5",
     "expo-constants": "~17.1.6",
     "expo-font": "~13.3.1",
     "expo-haptics": "~14.1.4",
-    "expo-image": "~2.1.7",
+    "expo-image": "~2.3.0",
     "expo-linking": "~7.1.5",
     "expo-random": "^14.0.1",
-    "expo-router": "~5.0.6",
-    "expo-splash-screen": "~0.30.8",
+    "expo-router": "~5.1.0",
+    "expo-splash-screen": "~0.30.9",
     "expo-status-bar": "~2.2.3",
-    "expo-symbols": "~0.4.4",
-    "expo-system-ui": "~5.0.7",
-    "expo-web-browser": "~14.1.6",
+    "expo-symbols": "~0.4.5",
+    "expo-system-ui": "~5.0.9",
+    "expo-web-browser": "~14.2.0",
     "firebase": "^11.8.1",
     "marked": "^15.0.12",
     "react": "19.0.0",
     "react-dom": "19.0.0",
-    "react-native": "0.79.2",
+    "react-native": "0.79.4",
     "react-native-draggable-flatlist": "^4.0.3",
     "react-native-gesture-handler": "~2.24.0",
     "react-native-markdown-display": "^7.0.2",
@@ -48,7 +49,7 @@
     "react-native-reanimated": "~3.17.4",
     "react-native-render-html": "^6.3.4",
     "react-native-safe-area-context": "5.4.0",
-    "react-native-screens": "~4.10.0",
+    "react-native-screens": "~4.11.1",
     "react-native-toast-message": "^2.3.0",
     "react-native-web": "~0.20.0",
     "react-native-webview": "13.13.5",
@@ -56,6 +57,7 @@
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
+    "@types/lodash.debounce": "^4.0.9",
     "@types/react": "~19.0.10",
     "eslint": "^9.25.0",
     "eslint-config-expo": "~9.2.0",
